chore: rename decorate to inspect
