+++ lib/index.js
@@ -1,11 +1,11 @@
 'use strict';
 
-const decorate = require('./decorate');
+const inspect = require('./inspect');
 
 module.exports = (log, levels = ['debug', 'info', 'warn', 'error']) => {
   const metadataLog = levels.reduce((proxy, level) => {
     proxy[level] = (msg, metadata) => {
-      return log[level](msg, metadata ? decorate(metadata) : undefined);
+      return log[level](msg, metadata ? inspect(metadata) : undefined);
     };
     return proxy;
   }, {});
