+++ lib/inspect.js
